Rating account crud

+++ src/main/java/com/soprasteria/devopsassesmenttool/controller/CategoryController.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,7 +15,6 @@ import org.springframework.web.bind.annotation.RestController;
 import com.soprasteria.devopsassesmenttool.model.Category;
 import com.soprasteria.devopsassesmenttool.sevice.CategoryService;
 import com.soprasteria.devopsassesmenttool.sevice.QuestionService;
-import com.soprasteria.devopsassesmenttool.util.CustomErrorType;
 
 @RestController
 @RequestMapping("/devops")
@@ -39,18 +37,18 @@ public class CategoryController {
 	}
 
 	@RequestMapping(value = "/category/{categoryId}", method = RequestMethod.GET)
-	public Optional<Category> getCategoryById(@PathVariable(value = "categoryId") Integer categoryId) {
+	public Optional<Category> getCategoryById(@PathVariable(value = "categoryId") Long categoryId) {
 		return categoryService.getCategoryById(categoryId);
 	}
 
 	@RequestMapping(value = "/category", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public Category updateCategory(@PathVariable(value = "categoryId") Integer categoryId,
+	public Category updateCategory(@PathVariable(value = "categoryId") Long categoryId,
 			@RequestBody Category category) {
 		return categoryService.updateCategoryById(categoryId, category);
 	}
 
 	@RequestMapping(value = "/category/{categoryId}", method = RequestMethod.DELETE)
-	public ResponseEntity<Object> deleteCategoryById(@PathVariable(value = "categoryId") Integer categoryId) {
+	public ResponseEntity<Object> deleteCategoryById(@PathVariable(value = "categoryId") Long categoryId) {
 		return categoryService.deleteCategoryById(categoryId);
 	}
