+++ src/main/java/com/soprasteria/devopsassesmenttool/model/Rating.java
@@ -5,7 +5,6 @@ package com.soprasteria.devopsassesmenttool.model;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -34,43 +33,32 @@ public class Rating implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	
-	private Integer rid;
+	private Long rid;
 
 	private Integer ratingValue;
 
 	private String ratinglabel;
 
 	private String ratingDesc;
-    
 
 	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "questionId", nullable = false)
 	@OnDelete(action = OnDeleteAction.CASCADE)
 	private Question question;
 
-
-
 	/**
 	 * @return the rid
 	 */
-	public Integer getRid() {
+	public Long getRid() {
 		return rid;
 	}
 
 	/**
 	 * @param rid the rid to set
 	 */
-	public void setRid(Integer rid) {
+	public void setRid(Long rid) {
 		this.rid = rid;
 	}
-
-
-
-	/**
-	 * @return the ratingValue
-	 */
-
 
 	/**
 	 * @return the ratingDesc
@@ -78,8 +66,6 @@ public class Rating implements Serializable {
 	public String getRatingDesc() {
 		return ratingDesc;
 	}
-
-
 
 	/**
 	 * @return the ratingValue
@@ -110,8 +96,7 @@ public class Rating implements Serializable {
 	}
 
 	/**
-	 * @param ratingDesc
-	 *            the ratingDesc to set
+	 * @param ratingDesc the ratingDesc to set
 	 */
 	public void setRatingDesc(String ratingDesc) {
 		this.ratingDesc = ratingDesc;
@@ -133,8 +118,7 @@ public class Rating implements Serializable {
 	}
 
 	/**
-	 * @param question
-	 *            the question to set
+	 * @param question the question to set
 	 */
 	public void setQuestion(Question question) {
 		this.question = question;
